+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Fixed
 * Fix delete button disabling after deletion on lists.
+* Fix dropdown's overflow on list components.
 * BlueSky theme:
     * Fix lookup dialog pager visibility.
     * Fix buttons imposition on each other in list's toolbar.